README: Fix inconsistency MF flag
The README said the AWUS036ACM could not properly inject frames with the More Fragment flag set. However, the paper says that the RT5572 chipsets had issues with this. In my experiment, the RT5572 (CSL dongle) had issues with the MF flag, where the workaround solved the issue. The AWUS036ACM was not tested.